@@ -1,1 +1,6 @@
-class NoParams {}
+import 'package:equatable/equatable.dart';
+
+class NoParams extends Equatable {
+  @override
+  List<Object> get props => [];
+}
